Moving service registration into the common spot.
