+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Services/ServiceRegistrations.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
+using Azure.AI.OpenAI;
 using Azure.Sdk.Tools.Cli.Helpers;
+using Microsoft.Extensions.Azure;
 
 namespace Azure.Sdk.Tools.Cli.Services
 {
@@ -26,6 +28,26 @@ namespace Azure.Sdk.Tools.Cli.Services
             services.AddSingleton<ITypeSpecHelper, TypeSpecHelper>();
             services.AddSingleton<ISpecPullRequestHelper, SpecPullRequestHelper>();
             services.AddSingleton<IUserHelper, UserHelper>();
+
+            services.AddAzureClients(clientBuilder =>
+            {
+                // For more information about this pattern: https://learn.microsoft.com/en-us/dotnet/azure/sdk/dependency-injection
+                var service = new AzureService();
+                clientBuilder.UseCredential(service.GetCredential());
+
+                // Azure OpenAI client does not, for some reason, have an
+                // in-package facade for this, so register manually.
+                clientBuilder.AddClient<AzureOpenAIClient, AzureOpenAIClientOptions>(
+                    (options, credential, _) =>
+                    {
+                        var endpointEnvVar = Environment.GetEnvironmentVariable("AZURE_OPENAI_ENDPOINT");
+                        var ep = string.IsNullOrWhiteSpace(endpointEnvVar) ?
+                            "https://openai-shared.openai.azure.com"
+                            : endpointEnvVar;
+
+                        return new AzureOpenAIClient(new Uri(ep), credential, options);
+                    });
+            });
         }
     }
 }
